Rework the isolated MSBuild-based project reading to a process
There were countless issues with the AppDomain-based isolation,
remoting, shared assemblies, propagating exceptions, binding
redirects from the parent process affecting indirect dependencies
of the child AppDomain (i.e. MSBuild 14 is an indirect reference
of the MSBuildWorkspace we need to use), so it was all getting
very convoluted.

To avoid all that, a new external process is provided that runs
in full isolation, from its app-local path where all the right
MSBuild and Roslyn dependencies are deployed (ILMerging may be
useful in the future to make it even more isolated from the
environment and just require .NET 4.5.2 installed).

The main ProjectLoader communicates with this external process
by directly writing to its input stream and reading from its
output stream, thus avoiding unnecessary file copying.

The process returns (as before) an XML structure with the basic
project, document and references metadata used to load the
workspace in the in-process task being run.

In addition, the extension method was moved to ITask, so that
the basic functionality can be used even from MSBuild 4.0
(even if to actually succeed in running, MSBuild 14 assemblies
will need to be present alongside the external project reader
process, until we figure out a way to ILMerge a .NET 4.5.2,
System.Runtime-based assembly).