[vsixinstaller] Don't uninstall VSIXes that will be upgraded
The VSIX APIs will already take care of upgrading properly. This makes for
faster execution, but also prevents potential failures when the VSIX being
upgraded is actually a built-in VS component that can't be uninstalled
ultimately, but it was found via the Exp inheriting the regular hive extensions.

This scenario works because VS will "override" the normal hive extension with
the newly installed Exp extension, so the "upgrade" (which will actually be
just a brand new install in this case) will just work.