Add "VSIX experimentalizer" behavior to VsixInstaller and VsixDependency
VSIXes that are "production" components cannot be simply installed:
they need to be made "Experimental" and only then you can freely
upgrade (and even downgrade) them.

Ths VsixExp.exe tool performs such operation, turning any production
VSIX into an experimental VSIX, just prior to installing it.

The VsixInstaller provides the path to $(VsixExp) which can later
on be used by the VsixDependency to experimentalize all @(VsixDependency)
declared in the project.

The VsixExp tool receives a SourceVsix and optional TargetVsix and writes
the former into the latter path (or in-place if not provided) the
experimental version of the VSIX (it it wasn't already), ready for installation.

If the received SourceVsix is already an experimental VSIX, it's just copied
over to the TargetVsix.