Get SDK-style GenerateAssemblyInfo on classic projects
This package imports the SDK-provided targets for assembly info
generation into a classic project. For an SDK-style project,
adding the package is a no-op, so it can safely be depended upon
by other packages.

The error message if the SDK is missing will also helpfully show
the entire command line to run to get a minimal install of it that
will satisfy the dependency for this package.